Move LinearMemory test from vm::Memory to Tunables
